+++ uspto/util/command.py
@@ -13,8 +13,8 @@ logger = logging.getLogger(__name__)
 def run_command(client, options):
     """
     Usage:
-      {program} get  <document-number> --type=publication --format=xml [--pretty] [--background] [--wait] [--debug]
-      {program} save <document-number> --type=publication --format=xml [--pretty] [--directory=/var/spool/uspto] [--use-application-id] [--overwrite] [--background] [--wait] [--debug]
+      {program} get  <document-number> --format=xml [--type=publication] [--pretty] [--background] [--wait] [--debug]
+      {program} save <document-number> --format=xml [--type=publication] [--pretty] [--directory=/var/spool/uspto] [--use-application-id] [--overwrite] [--background] [--wait] [--debug]
       {program} bulk get  --numberfile=numbers.txt --format=xml,json [--pretty] [--wait] [--debug]
       {program} bulk save --numberfile=numbers.txt --format=xml,json [--pretty] --directory=/var/spool/uspto [--use-application-id] [--overwrite] [--wait] [--debug]
       {program} info
@@ -22,16 +22,22 @@ def run_command(client, options):
       {program} (-h | --help)
 
     General options:
-      --type=<type>             Document type, one of "publication", "application" or "patent".
+      <document-number>         Document number, e.g. 2017/0293197, US20170293197A1, PP28532, 15431686.
+                                Format depends on data source.
+      --type=<type>             Document type, one of "publication", "application", "patent" or "auto".
+
+    Output options:
       --format=<target>         Data format, one of "xml" or "json".
       --pretty                  Pretty-print output data. Currently applies to "--format=json" only.
-      --background              Run the download process in the background.
-      --wait                    Wait for the background download to finish.
 
     Save options:
       --directory=<directory>   Save downloaded to documents to designated target directory.
       --use-application-id      When saving documents, use the application identifier as filename.
       --overwrite               When saving documents, overwrite already existing documents.
+
+    Operation mode:
+      --background              Run the download process in the background.
+      --wait                    Wait for the background download to finish.
 
     Bulk options:
       --numberfile=<numberfile> Read document numbers from file.